DM-6726 fixed expanded mode and charttest demo,  now using default parameters in app mode (when no xCol and yCol are present)
