+++ src/firefly/js/visualize/ChartUtil.js
@@ -11,6 +11,7 @@ import {get} from 'lodash';
 import {uniqueId} from 'lodash';
 
 import {flux} from '../Firefly.js';
+import {MetaConst} from '../data/MetaConst.js';
 import {getTblById, getColumnIdx, getCellValue} from '../tables/TableUtil.js';
 import {Expression} from '../util/expr/Expression.js';
 import {logError} from '../util/WebUtil.js';
@@ -95,6 +96,62 @@ export function numRelatedCharts(tblId) {
 
 export function uniqueChartId(prefix) {
     return uniqueId(prefix?prefix+'-c':'c');
+}
+
+function colWithName(cols, name) {
+    return cols.find((c) => { return c.name===name; });
+}
+
+function getNumericCols(cols) {
+    const ncols = [];
+    cols.forEach((c) => {
+        if (c.type !== 'char') {
+            ncols.push(c);
+        }
+    });
+    return ncols;
+}
+
+export function getDefaultXYPlotParams(tbl_id) {
+
+    const {tableMeta, tableData, totalRows}= getTblById(tbl_id);
+
+    if (!totalRows) {
+        return;
+    }
+
+    // for catalogs use lon and lat columns
+    let isCatalog = Boolean(tableMeta[MetaConst.CATALOG_OVERLAY_TYPE] && tableMeta[MetaConst.CATALOG_COORD_COLS]);
+    let xCol = undefined, yCol = undefined;
+
+    if (isCatalog) {
+        const s = tableMeta[MetaConst.CATALOG_COORD_COLS].split(';');
+        if (s.length !== 3) return;
+        xCol = colWithName(tableData.columns, s[0]); // longtitude
+        yCol = colWithName(tableData.columns, s[1]); // latitude
+
+        if (!xCol || !yCol) {
+            isCatalog = false;
+        }
+    }
+
+    // otherwise use the first one-two numeric columns
+    if (!isCatalog) {
+        const numericCols = getNumericCols(tableData.columns);
+        if (numericCols.length > 2) {
+            xCol = numericCols[0];
+            yCol = numericCols[1];
+        } else if (numericCols.length > 1) {
+            xCol = numericCols[0];
+            yCol = numericCols[0];
+        }
+    }
+
+    return (xCol && yCol) ?
+    {
+        x: {columnOrExpr: xCol.name, label: xCol.name, unit: xCol.units?xCol.units:''},
+        y: {columnOrExpr: yCol.name, label: yCol.name, unit: yCol.units?yCol.units:''}
+    } : undefined;
 }
