+++ src/firefly/js/visualize/saga/ChartsSync.js
@@ -6,7 +6,6 @@ import {take} from 'redux-saga/effects';
 import {has} from 'lodash';
 
 import {flux} from '../../Firefly.js';
-import {MetaConst} from '../../data/MetaConst.js';
 
 import * as TablesCntlr from '../../tables/TablesCntlr.js';
 import * as TableStatsCntlr from '../TableStatsCntlr.js';
@@ -14,9 +13,7 @@ import * as TableUtil from '../../tables/TableUtil.js';
 
 import * as XYPlotCntlr from '../XYPlotCntlr.js';
 import * as HistogramCntlr from '../HistogramCntlr.js';
-import {hasRelatedCharts} from '../ChartUtil.js';
-
-
+import {hasRelatedCharts, getDefaultXYPlotParams} from '../ChartUtil.js';
 
 /**
  * this saga handles chart related side effects
@@ -84,58 +81,5 @@ export function* syncCharts() {
     }
 }
 
-function colWithName(cols, name) {
-    return cols.find((c) => { return c.name===name; });
-}
 
-function getNumericCols(cols) {
-    const ncols = [];
-    cols.forEach((c) => {
-        if (c.type !== 'char') {
-            ncols.push(c);
-        }
-    });
-    return ncols;
-}
 
-function getDefaultXYPlotParams(tbl_id) {
-
-    const {tableMeta, tableData, totalRows}= TableUtil.getTblById(tbl_id);
-
-    if (!totalRows) {
-        return;
-    }
-
-    // for catalogs use lon and lat columns
-    let isCatalog = Boolean(tableMeta[MetaConst.CATALOG_OVERLAY_TYPE] && tableMeta[MetaConst.CATALOG_COORD_COLS]);
-    let xCol = undefined, yCol = undefined;
-
-    if (isCatalog) {
-        const s = tableMeta[MetaConst.CATALOG_COORD_COLS].split(';');
-        if (s.length !== 3) return;
-        xCol = colWithName(tableData.columns, s[0]); // longtitude
-        yCol = colWithName(tableData.columns, s[1]); // latitude
-
-        if (!xCol || !yCol) {
-            isCatalog = false;
-        }
-    }
-
-    // otherwise use the first one-two numeric columns
-    if (!isCatalog) {
-        const numericCols = getNumericCols(tableData.columns);
-        if (numericCols.length > 2) {
-            xCol = numericCols[0];
-            yCol = numericCols[1];
-        } else if (numericCols.length > 1) {
-            xCol = numericCols[0];
-            yCol = numericCols[0];
-        }
-    }
-
-    return (xCol && yCol) ?
-    {
-        x: {columnOrExpr: xCol.name, label: xCol.name, unit: xCol.units?xCol.units:''},
-        y: {columnOrExpr: yCol.name, label: yCol.name, unit: yCol.units?yCol.units:''}
-    } : undefined;
-}
